More work on the allocations

+++ core_perks/memory/arena_allocator.h
@@ -11,9 +11,10 @@ namespace cp
 		void* alloc(size_t size, size_t alignment = alignof(std::max_align_t));
 		void free(void*) {}
 		void reset() { offset_ = 0; }
+
 	private:
 		uint8* memory_ = nullptr;
-		size_t offset_ = 0;
-		size_t capacity_ = 0;
+		uint64 offset_ = 0;
+		uint64 capacity_ = 0;
 	};
 }
